Update du lieu moi
